homework/requirements.txt
diskcache==5.6.3
Jinja2==3.1.6
llama_cpp_python==0.3.20
MarkupSafe==3.0.3
numpy==2.4.4
typing_extensions==4.15.0
